requirements.txt
ipywidgets
datasets
matplotlib
tensorflow-datasets
torchtext == 0.4

ipykernel == 6.29.0
torch == 2.1.2
torchsummary == 1.5.1
torchvision == 0.16.2
transformers == 4.36.2
tensorflow-cpu == 2.15.0